GS-14926 Displaying topical reviews on profile pages GS-14926 hide star rating, fix reporting content GS-14926 Reporting a review is sticky GS-14926 Filter by poster type GS-14926 Pagination GS-14926 Fix unit tests prior to check in

+++ src/java/gs/web/school/review/SchoolReviewsAjaxController.java
@@ -179,7 +179,7 @@ public class SchoolReviewsAjaxController extends AbstractCommandController imple
             }
         }
 
-        Long existingReviewId = review.getId();
+        Integer existingReviewId = review.getId();
         Poster poster = reviewCommand.getPoster();
 
         StringBuilder reasonToReportReview;
@@ -258,7 +258,7 @@ public class SchoolReviewsAjaxController extends AbstractCommandController imple
         }
 
         if (reasonToReportReview != null) {
-            getReportContentService().reportContent(getAlertWordFilterUser(), user, request, review.getId().intValue(), ReportedEntity.ReportedEntityType.topicalSchoolReview, reasonToReportReview.toString());
+            getReportContentService().reportContent(getAlertWordFilterUser(), user, request, review.getId(), ReportedEntity.ReportedEntityType.topicalSchoolReview, reasonToReportReview.toString());
         }
 
         //trigger the success events
